more code
added the hearts and stuff and made it work with a couple of changes. also added the score and high score which is saved into files. also added a menu and stuff. haven't done this in a while so a lot of features at once.lastly deletes unnecessary sprites and obects.